+++ internal/mobile/settings.go
@@ -1,7 +1,0 @@
-package mobile
-
-const (
-	// CAndroidRootPath = "." // only for tests
-	CAndroidRootPath = "/sdcard/Documents"
-	CAndroidFullPath = CAndroidRootPath + "/hidden_lake"
-)